+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Added quick accessor for is_invalidated and set_invalidated.
 - Renamed `set_invalidate` to `set_invalidated` in config flag.
 - Fixed button logic so that callback only occurs when mouse released over the initiated object (#245)
+- Modified `HorizontalLayout` and `VerticalLayout` to respect full padding (#247)
+- Modified layout test so that it contains horizontal and vertical layouts.
 
 ## 0.4.16